isuue 49 get focus on the search text

+++ src/widget/unit_control_panel/imp.rs
@@ -291,7 +291,7 @@ impl UnitControlPanelImpl {
                         }
                     }
                 })
-                .parameter_type(Some(glib::VariantTy::BYTE))
+                .parameter_type(Some(glib::VariantTy::INT32))
                 .build()
         };
